content/media/webaudio/test/test_delayNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test DelayNode</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script src="webaudio.js" type="text/javascript"></script>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);

  var context = new AudioContext();
  var buffer = context.createBuffer(1, 2048, context.sampleRate);
  for (var i = 0; i < 2048; ++i) {
    buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
  }
  var expectedBuffer = context.createBuffer(1, 2048 * 2, context.sampleRate);
  for (var i = 2048; i < 2048 * 2; ++i) {
    expectedBuffer.getChannelData(0)[i] = buffer.getChannelData(0)[i - 2048];
  }

  var destination = context.destination;

  var source = context.createBufferSource();

  var delay = context.createDelay();

  var delay2 = context.createDelayNode();
  isnot(delay, delay2, "createDelayNode should create a different delay node");

  source.buffer = buffer;

  source.connect(delay);
  var sp = context.createScriptProcessor(2048 * 2, 1);
  delay.connect(sp);
  sp.connect(destination);

  ok(delay.delayTime, "The audioparam member must exist");
  is(delay.delayTime.value, 0, "Correct initial value");
  is(delay.delayTime.defaultValue, 0, "Correct default value");
  delay.delayTime.value = 0.5;
  is(delay.delayTime.value, 0.5, "Correct initial value");
  is(delay.delayTime.defaultValue, 0, "Correct default value");

  var delay2 = context.createDelay(2);
  is(delay2.delayTime.value, 0, "Correct initial value");
  is(delay2.delayTime.defaultValue, 0, "Correct default value");
  delay2.delayTime.value = 0.5;
  is(delay2.delayTime.value, 0.5, "Correct initial value");
  is(delay2.delayTime.defaultValue, 0, "Correct default value");

  expectException(function() {
    context.createDelay(0);
  }, DOMException.NOT_SUPPORTED_ERR);
  expectException(function() {
    context.createDelay(180);
  }, DOMException.NOT_SUPPORTED_ERR);
  expectTypeError(function() {
    context.createDelay(NaN);
  }, DOMException.NOT_SUPPORTED_ERR);
  expectException(function() {
    context.createDelay(-1);
  }, DOMException.NOT_SUPPORTED_ERR);
  context.createDelay(1); // should not throw

  // Delay the source stream by 2048 frames
  delay.delayTime.value = 2048 / context.sampleRate;

  source.start(0);
  sp.onaudioprocess = function(e) {
    is(e.inputBuffer.numberOfChannels, 1, "Correct input channel count");
    compareBuffers(e.inputBuffer.getChannelData(0), expectedBuffer.getChannelData(0));

    sp.onaudioprocess = null;

    SpecialPowers.clearUserPref("media.webaudio.enabled");
    SimpleTest.finish();
  };
});

</script>
</pre>
</body>
</html>
